switch from junit to assertJ assertions

+++ src/test/java/api/pojo/response/beer/AddBeerResponse.java
@@ -8,7 +8,7 @@ public record AddBeerResponse(String message, BeerDetails beer) {
             String id,
             String name,
             String style,
-            int breweryId,
+            long breweryId,
             String ounces
     ) {
     }
